1. 创建订单时, "订单数量" 放在"我司料名"下面,属跟单人员填写,电镀要求,特殊要求,放在"半成品数"下面,属跟单人员填写. 2. 生产记录,删除他司料名 3. 显示未完成的生产纪录 4. 生产记录,移置顶端菜单,人人可见. 5. 生产记录,我司料名,引用数据库错误 6.新建顶部菜单项 "生产管理", 里面包括功能: A查看订单(内容与"订单管理"里的"查看订单"一样,但删掉"客户名称"与"他司料名". B审核订单(将"订单管理"里的"审核订单"移到这里) 7.在"生产管理"的"查看订单"里,修改只能修改下面部分,上面部分显灰,但可以修改订单数量 8.在"订单管理"的"查看订单"里,修改只能修改上面部分,但不能修改订单数量,下面部分显灰,并去掉或没有权限"补数"与"减数" 9.需要新增或修改"厂长,经理,生产经理"的权限, 使一个权限不能去"订单管理"与"后台管理",可以去"生产管理","仓库管理","文档管理",   并且这个权限改名为"生产部" 10.在"订单管理"里, 当生产已经开始进行(已经在某部门有完成数),则不能"取消订单与删除订单" 11. 在所有列表里, 添加 筛选功能, 比如某跟单人员,在订单里,筛选跟单人员"A", 就能只看自己follow的订单. 12. 添加搜索功能 13. 系统角色改成所属部门 14. 部门添加多一个"数控" 15. 标题改为奇迪科管理系统 16. TOTAL,改为总计

+++ ohtune/src/tk/solaapps/ohtune/controller/OrderController.java
@@ -326,8 +326,9 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 		String sRequirement3 = request.getParameter("requirement3");
 		String sRequirement4 = request.getParameter("requirement4");
 		String sCDeadline = request.getParameter("c_deadline");
+		String sEQuantity = "0" + request.getParameter("e_quantity");
 		
-		int iUseFinished,iUseSemiFinished;
+		int iUseFinished,iUseSemiFinished, iEQuantity;
 		
 		Enumeration<String> parms = request.getParameterNames();
 		
@@ -341,9 +342,10 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 		{
 			iUseFinished = Integer.parseInt(sUseFinished);
 			iUseSemiFinished = Integer.parseInt(sUseSemiFinished);
+			iEQuantity = Integer.parseInt(sEQuantity);
 		}catch(Exception e)
 		{
-			jr = service.genJsonResponse(false, "部门分配或成品分配数据错误", null);
+			jr = service.genJsonResponse(false, "订单数量成品半成品数量错误", null);
 			response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
 			return;
 		}
@@ -410,8 +412,8 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 		order.setRequirement_4(sRequirement4);
 		order.setPriority(0);
 		order.setQuantity(0);
-		order.setE_quantity(0);
-		order.setProduct_rate(1f);
+		order.setE_quantity(iEQuantity);
+		order.setProduct_rate(0f);
 		
 		List<JobType> jobTypes = new ArrayList<JobType>();
 		List<Job> jobs = new ArrayList<Job>();
@@ -459,14 +461,13 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 		String sQuantity1 = "0" + request.getParameter("quantity1");
 		String sQuantity2 = "0" + request.getParameter("quantity2");
 		String sQuantity3 = "0" + request.getParameter("quantity3");
-		String sEQuantity = "0" + request.getParameter("e_quantity");
 		String sQuantity = "0" + request.getParameter("quantity");
 		String sProductRate = request.getParameter("product_rate");
 		String sOrderId = request.getParameter("orderid");
 		String sPriority = request.getParameter("priority");
 		String sDeadline = request.getParameter("deadline");
 		
-		int iQuantity,iQuantity1,iQuantity2,iQuantity3,iUseFinished,iUseSemiFinished,iEQuantity;
+		int iQuantity,iQuantity1,iQuantity2,iQuantity3,iUseFinished,iUseSemiFinished;
 		float fProductRate;
 		
 		Gson gson = service.getGson();
@@ -493,7 +494,6 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 			iQuantity1 = Integer.parseInt(sQuantity1);
 			iQuantity2 = Integer.parseInt(sQuantity2);
 			iQuantity3 = Integer.parseInt(sQuantity3);
-			iEQuantity = Integer.parseInt(sEQuantity);
 		}catch(Exception e)
 		{
 			jr = service.genJsonResponse(false, "部门分配或成品分配数据错误", null);
@@ -550,7 +550,6 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 		
 		order.setQuantity(iQuantity1 + iQuantity2 + iQuantity3);
 		order.setProduct_rate(fProductRate);
-		order.setE_quantity(iEQuantity);
 		order.setQuantity(iQuantity);
 		
 		if(iQuantity != iQuantity1 + iQuantity2 + iQuantity3)
@@ -652,11 +651,19 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 			}
 			else
 			{
-				boolean success = service.deleteJobByOrder(order, sessionUser);
-				if(success)
-					success = success & service.deleteOrder(order);
-				jr = service.genJsonResponse(success, success ? "删除订单成功" : "删除订单失败", null);
-				response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
+				if(order.getStatus().equals(Order.STATUS_APPROVING) || order.getStatus().equals(Order.STATUS_FINISHED))
+				{
+					boolean success = service.deleteJobByOrder(order, sessionUser);
+					if(success)
+						success = success & service.deleteOrder(order);
+					jr = service.genJsonResponse(success, success ? "删除订单成功" : "删除订单失败", null);
+					response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
+				}
+				else
+				{
+					jr = service.genJsonResponse(false, "删除订单失败， 订单已经在进行中", null);
+					response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
+				}
 			}
 		}
 		catch(Exception e)
@@ -766,9 +773,16 @@ public class OrderController extends HttpServlet implements IOhtuneController{
 			}
 			else
 			{
-				boolean success = service.cancelOrder(order, sessionUser);
-				jr = service.genJsonResponse(success, success ? "取消订单成功" : "取消订单失败", null);
-				response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
+				if(order.getStatus().equals(Order.STATUS_APPROVING) || order.getStatus().equals(Order.STATUS_FINISHED))
+				{
+					boolean success = service.cancelOrder(order, sessionUser);
+					jr = service.genJsonResponse(success, success ? "取消订单成功" : "取消订单失败", null);
+					response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
+				}else
+				{
+					jr = service.genJsonResponse(false, "取消订单失败， 订单已经在进行中", null);
+					response.getOutputStream().write(gson.toJson(jr).getBytes("utf-8"));
+				}
 			}
 		}
 		catch(Exception e)
